# swagger.json
{
    "openapi":  "3.0.1",
    "info":  {
                 "title":  "FinVeda API",
                 "version":  "v1"
             },
    "paths":  {
...
                  "/api/v1/Installments/generate/{loanId}":  {
                                                                 "post":  {
                                                                              "tags":  [
                                                                                           "Installments"
                                                                                       ],
                                                                              "parameters":  [
                                                                                                 {
                                                                                                     "name":  "loanId",
                                                                                                     "in":  "path",
                                                                                                     "required":  true,
                                                                                                     "schema":  "@{type=string; format=uuid}"
                                                                                                 },
                                                                                                 {
                                                                                                     "name":  "count",
                                                                                                     "in":  "query",
                                                                                                     "schema":  "@{type=integer; format=int32; default=12}"
                                                                                                 }
                                                                                             ],
                                                                              "responses":  {
                                                                                                "200":  {
                                                                                                            "description":  "OK"
                                                                                                        }
                                                                                            }
                                                                          }
                                                             },
                  "/api/v1/Installments/due":  {
                                                   "get":  {
                                                               "tags":  [
                                                                            "Installments"
                                                                        ],
                                                               "parameters":  [
                                                                                  {
                                                                                      "name":  "from",
                                                                                      "in":  "query",
                                                                                      "schema":  "@{type=string; format=date-time}"
                                                                                  },
                                                                                  {
                                                                                      "name":  "to",
                                                                                      "in":  "query",
                                                                                      "schema":  "@{type=string; format=date-time}"
                                                                                  }
                                                                              ],
                                                               "responses":  {
                                                                                 "200":  {
                                                                                             "description":  "OK",
                                                                                             "content":  "@{text/plain=; application/json=; text/json=}"
                                                                                         }
                                                                             }
                                                           }
                                               },
                  "/api/v1/Installments/loan/{loanId}":  {
                                                             "get":  {
                                                                         "tags":  [
                                                                                      "Installments"
                                                                                  ],
                                                                         "parameters":  [
                                                                                            {
                                                                                                "name":  "loanId",
                                                                                                "in":  "path",
                                                                                                "required":  true,
                                                                                                "schema":  "@{type=string; format=uuid}"
                                                                                            }
                                                                                        ],
                                                                         "responses":  {
                                                                                           "200":  {
                                                                                                       "description":  "OK",
                                                                                                       "content":  "@{text/plain=; application/json=; text/json=}"
                                                                                                   }
                                                                                       }
                                                                     }
                                                         },
...
              },
    "components":  {
...
                       "securitySchemes":  {
                                               "Bearer":  {
                                                              "type":  "http",
                                                              "description":  "Enter \u0027Bearer\u0027 [space] and then your valid token.",
                                                              "scheme":  "Bearer",
                                                              "bearerFormat":  "JWT"
                                                          }
                                           }
                   },
    "security":  [
                     {
                         "Bearer":  [

                                    ]
                     }
                 ]
}